Leonardo

+++ src/map/MapSensor.h
@@ -5,10 +5,10 @@
 class MapSensor {
 public:
     MapSensor(uint8_t pin);
-    int getRead() { return analogRead(PIN_MAP); }
-    float getVoltage() { return (float)getRead() * 5.f / 1023.f; }
-    int getPressure() { return turnVoltageToPressure(getVoltage()); }
+    uint32_t getRead() { return analogRead(PIN_MAP); }
+    double getVoltage() { return (double)getRead() * 5.f / 1023.f; }
+    uint32_t getPressure() { return turnVoltageToPressure(getVoltage()); }
 private:
-    int turnVoltageToPressure(float voltage) { return round(21.25 * voltage + 4.375); }
+    uint32_t turnVoltageToPressure(double voltage) { return round(21.25 * voltage + 4.375); }
     uint8_t PIN_MAP;
 };
